feat: moving committed files to dedicated namespace.
